requirements.txt
# PaddleOCR Server Dependencies
# Matches local Windows environment versions

paddlepaddle-gpu==3.2.0
paddleocr==3.2.0
numpy==2.2.6
opencv-python-headless==4.10.0.84
pydantic==2.12.0
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
